Rename FileResponse interface to BoreholeFile
